Enabled removal of original
Added configuration option REMOVE_ORIGINAL, which if set to true
will remove the original recording.  This only is checked if
UPDATE_DATABASE is set to true as well.  If the original is not
to be removed, it is renamed with a '.old' extension to clearly
mark it in the recordings directory.